Added reaction time analysis and dotplot. Also added connection between frontend data to backend data

+++ backend/PassionProjectReactionTimeBackend/src/main/java/com/WayneRagsac/PassionProjectReactionTimeBackend/services/ReactionTimeService.java
@@ -43,7 +43,7 @@ public class ReactionTimeService implements ReactionTimeServiceInterface {
     }
 
     @Override
-    public ReactionTimeEntity addReactionTime(String userId, long reactionTime) {
+    public ReactionTimeEntity addReactionTime(String userId, int reactionTime) {
         try {
             ReactionTimeEntity newReactionTime = new ReactionTimeEntity();
             newReactionTime.setUserId(userId);
